Continue to add foreign key references
Small rewrite to move commas to EOL to match SQL conventions
Move purchases migration to be after grants since it references grants
table

+++ db/sql/004_create_students.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS tbl_fact_student(
-        uin smallint UNSIGNED NOT NULL auto_increment PRIMARY KEY,
+        uin int UNSIGNED NOT NULL auto_increment PRIMARY KEY,
         name varchar(50),
         department_id int UNSIGNED,
         program ENUM('UG', 'MS', 'PHD'),
